chore(rename-folder): changed utils folder to helpers
