+++ src/helpers/formatStrings.ts
