+++ src/pages/CompanyListPage.tsx
@@ -16,7 +16,7 @@ import { CustomExpandableTable } from '../components/CustomExpandableTable';
 import SearchCompanyById from '../components/SearchCompanyById';
 
 import { Company } from '../@types/CompanyTypes';
-import { toBrasiliaTime } from '../utils/dateFormater';
+import { toBrasiliaTime } from '../helpers/dateFormater';
 import { useCompanyStore } from '../stores/company.store';
 import { REMOVE_CONFIRM, REMOVE_CONTENT } from '../constants/dialogMessages';
